+++ Tests/Publish.CleanArch.MassTransit.OutboxNone.TestApplication/Publish.CleanArch.MassTransit.OutboxNone.TestApplication.Infrastructure/Configuration/MassTransitConfiguration.cs
@@ -3,6 +3,7 @@ using System.Reflection;
 using Intent.RoslynWeaver.Attributes;
 using MassTransit;
 using MassTransit.Configuration;
+using MassTransit.Messages.Shared;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -30,8 +31,14 @@ namespace Publish.CleanArch.MassTransit.OutboxNone.TestApplication.Infrastructur
                         host.Password(configuration["RabbitMq:Password"]);
                     });
                     cfg.ConfigureEndpoints(context);
+                    cfg.AddMessageTopologyConfiguration();
                 });
             });
+        }
+
+        private static void AddMessageTopologyConfiguration(this IRabbitMqBusFactoryConfigurator cfg)
+        {
+            cfg.Message<BasketCreatedEvent>(x => x.SetEntityName("basket-created-topic-rename"));
         }
 
         private static void AddConsumers(this IRegistrationConfigurator cfg)